Move new udpate notification
Currently the new update notification if an announcement is less than seven days,
is on the trailing end of the list widget. While this works perfectly on lager devices,
the view is skewed on smalled mobile devices.

To fix the issue, the trailing update is removed and instead the listview is wrapped in a
Column which has a Text on top of the listview card with 'NEW UPDATE'.